adjust for install sh

+++ install.sh
@@ -31,7 +31,7 @@ else
     echo "You have installed node"
 fi
 
-if [[ ! -e /usr/bin/gsed ]]; then
+if [[ ! -e /usr/local/bin/gsed ]]; then
     brew install gnu-sed
 else
     echo "You have installed gsed"
@@ -43,26 +43,40 @@ else
     echo "You have installed ctags"
 fi
 
+if [[ ! -e /usr/local/bin/pod ]]; then
+    sudo gem install cocoapods
+else
+    echo "You have installed cocoapods"
+fi
+if [[ ! -e /usr/local/bin/fastlane ]]; then
+    sudo gem install colored
+    sudo gem install -n /usr/local/bin fastlane
+else
+    echo "You have installed fastlane"
+fi
+
+
 if [[ ! -e ~/.oh-my-zsh ]]; then
     sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+else
+    echo "You have downloaded oh my zsh"
 fi
 
 if [[ ! -e ~/Library/Developer/Xcode/UserData/CodeSnippets ]]; then
     git clone https://github.com/hanl001/CodeSnippets.git ~/Library/Developer/Xcode/UserData/CodeSnippets
+else
+    echo "You have cloned xcode code snippets"
 fi
 
-if [[ ! -e "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions" ]]; then
+if [[ ! -e ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions ]]; then
     git clone git@github.com:zsh-users/zsh-autosuggestions.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+else
+    echo "You have cloned zsh autosuggestions"
 fi
 
-if [[ ! -e "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting" ]]; then
+if [[ ! -e ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting ]]; then
     git clone git@github.com:zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+else
+    echo "You have cloned zsh syntax highlighting"
 fi
-
-# Gem update
-sudo gem update --system
-sudo gem install -n /usr/local/bin fastlane
-sudo gem install cocoapods -v 1.9.3 
-sudo gem install colored
-
